fixed bad commit

+++ include/tokenizer.h
@@ -74,6 +74,6 @@ bool is_number(char c);
 bool is_alpha(char c);
 void print_token(Token tok);
 TokenArray tokenize(const char* buf, size_t buf_size);
-KeywordEntry *lookup_keyword(const char *str, size_t len);
+KeywordEntry *lookup_keyword(const char *str, unsigned int len); // I think this should be an unsigned int, as that is what GPERF generates
 
 #endif
